Remove XGAP_DIR and directly set XGAP variable
It is completely unclear to me what the point of XGAP_DIR
was; there are some mentions of "xgap libraries", but I couldn't
find anything that suggested one has to set this "XGAP_DIR" as
an additional GAP root.

So let's ditch it, which allows us to fix the XGAP value
by setting it directly, based on the $srcdir.

Fixes #2