create xgap_test.g, a framework to test GAP package mode
This framework loads a new instance of GAP run with the -p flag, parses and responds to XGAP queries about fonts and colors, and waits until GAP is ready for input.  It then runs a single command, waits until GAP is ready for input (or error input) again, and then closes the child GAP -p instance.

All output from the GAP -p instance goes to stdout.  The package mode codes are parse and properly displayed: the output should look like that of a normal GAP instance.