Fix out-of-bounds read & write
gcc warns:

selfile.c:1253:24: warning: "strncat" specified bound depends on the
length of the source argument [-Wstringop-overflow=]
 1253 |                 (void) strncat(SFcurrentPath, str, len - 1);

The third argument ought to be the size of the _destination_, not the
_source_. You need snprintf (or a contrived mix of min()+max()) if
you want to also restrict the source length.

Also, `len - 1` can underflow if the string length for some
reason is 0, catch that case as well.