Resolve build failure under gcc 15/-std=c23
```
../../src.x11/gaptext.c: In function ‘DoSelection’:
../../src.x11/gaptext.c:156:13: error: too many arguments to function ‘gap->gap.input_callback’; expected 0, have 2
 156 |             gap->gap.input_callback( buf, 2 );
../../src.x11/gaptext.h:99:27: note: declared here
   99 |     void                (*input_callback)();
../../src.x11/gaptext.c: In function ‘GapTextSelectStart’:
../../src.x11/gaptext.c:341:25: error: too many arguments to function ‘gap->gap.check_caret_pos’; expected 0, have 2
  341 |         newPos = oldPos+gap->gap.check_caret_pos(gap->text.insertPos,oldPos);
../../src.x11/gaptext.h:102:27: note: declared here
  102 |     Int                 (*check_caret_pos)();
(and others that are more obvious)
```

It is unclear how x.input_callback and x.check_caret_pos ever get
assigned. Due to the use of unspecified argument lists previously,
`check_caret_pos` would be called with arguments of type
XawTextPosition (which are `long`!) whereas `input_callback` would be
called with an `int` (which is not so long).

It is conceivable that the actual function behind some
`input_callback` pointer is expecting XawTextPosition and, because
the function is not getting it, there has been (and continues to do
with this change) illegal memory access past the end of the
arguments-on-stack.